New changes

+++ database/migrations/2021_01_11_133526_create_social_instagram_accounts_table.php
@@ -14,7 +14,7 @@ class CreateSocialInstagramAccountsTable extends Migration
     public function up()
     {
         Schema::create('social_instagram_accounts', function (Blueprint $table) {
-            $table->integer('user_id');
+            $table->bigIncrements('user_id');
             $table->string('provider_user_id');
             $table->string('provider');
             $table->timestamps();
